requirements.txt
# 📦 DEPENDENCIES SMART WASTE CLASSIFIER
# Install dengan: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Machine Learning
tensorflow>=2.13.0
keras>=2.13.0
scikit-learn>=1.3.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Data Handling
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
